Addition of business logic in each processor, according to specified rules.

+++ src/main/java/com/adidas/pac/model/input/PaymentRules.java
@@ -9,5 +9,5 @@ import lombok.Value;
 @Value
 public class PaymentRules {
   @JsonProperty("max-limit")
-  private Integer maxLimit;
+  private Long maxLimit;
 }
